+++ src/main/java/com/example/healthgenie/boundedContext/community/controller/CommunityPostController.java
@@ -1,20 +1,17 @@
 package com.example.healthgenie.boundedContext.community.controller;
 
-import com.example.healthgenie.boundedContext.community.dto.PostRequest;
-import com.example.healthgenie.boundedContext.community.dto.PostResponse;
 import com.example.healthgenie.base.exception.CommunityPostException;
 import com.example.healthgenie.base.utils.SecurityUtils;
-import com.example.healthgenie.boundedContext.community.service.CommunityPostPhotoService;
+import com.example.healthgenie.boundedContext.community.dto.PostRequest;
+import com.example.healthgenie.boundedContext.community.dto.PostResponse;
 import com.example.healthgenie.boundedContext.community.service.CommunityPostService;
-import com.example.healthgenie.base.utils.S3UploadUtils;
+import com.example.healthgenie.boundedContext.community.service.CommunityPostTransactionService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
@@ -27,8 +24,7 @@ import static com.example.healthgenie.base.exception.CommunityPostErrorResult.NO
 public class CommunityPostController {
 
     private final CommunityPostService communityPostService;
-    private final CommunityPostPhotoService communityPostPhotoService;
-    private final S3UploadUtils s3UploadUtils;
+    private final CommunityPostTransactionService communityPostTransactionService;
 
     @GetMapping("/{id}")
     public ResponseEntity<PostResponse> findById(@PathVariable Long id) {
@@ -37,27 +33,7 @@ public class CommunityPostController {
 
     @PostMapping
     public ResponseEntity<PostResponse> save(PostRequest request) throws IOException {
-        // 이미지 S3 저장
-        List<String> photoPaths = new ArrayList<>();
-        if(existsFile(request)) {
-            photoPaths = s3UploadUtils.upload(request.getPhotos(), "post-photos");
-        }
-
-        // CommunityPost 엔티티 저장
-        PostResponse savedPost = communityPostService.save(request);
-
-        // CommunityPostPhoto 엔티티 저장
-        if(existsFile(request)) {
-            communityPostPhotoService.saveAll(savedPost.getId(), photoPaths);
-        }
-
-        PostResponse response = PostResponse.builder()
-                .id(savedPost.getId())
-                .title(savedPost.getTitle())
-                .content(savedPost.getContent())
-                .userId(savedPost.getUserId())
-                .photoPaths(photoPaths)
-                .build();
+        PostResponse response = communityPostTransactionService.save(request);
 
         return ResponseEntity.ok(response);
     }
@@ -69,27 +45,7 @@ public class CommunityPostController {
             throw new CommunityPostException(NO_PERMISSION);
         }
 
-        // 이미지 S3 저장
-        List<String> photoPaths = new ArrayList<>();
-        if(existsFile(request)) {
-            photoPaths = s3UploadUtils.upload(request.getPhotos(), "post-photos");
-        }
-
-        // CommunityPost 엔티티 저장
-        PostResponse updatedPost = communityPostService.update(id, request);
-
-        // CommunityPostPhoto 엔티티 저장
-        if (existsFile(request)) {
-            communityPostPhotoService.updateAll(updatedPost.getId(), photoPaths);
-        }
-
-        PostResponse response = PostResponse.builder()
-                .id(updatedPost.getId())
-                .title(updatedPost.getTitle())
-                .content(updatedPost.getContent())
-                .userId(updatedPost.getUserId())
-                .photoPaths(photoPaths)
-                .build();
+        PostResponse response = communityPostTransactionService.update(id, request);
 
         return ResponseEntity.ok(response);
     }
@@ -103,15 +59,7 @@ public class CommunityPostController {
 
         communityPostService.delete(id);
 
-        return ResponseEntity.ok(id + "번 게시글이 삭제되었습니다.");
-    }
-
-    private boolean existsFile(PostRequest request) {
-        long totalFileSize = request.getPhotos().stream()
-                .mapToLong(MultipartFile::getSize)
-                .sum();
-
-        return !request.getPhotos().isEmpty() && totalFileSize > 0;
+        return ResponseEntity.ok("게시글이 삭제되었습니다.");
     }
 
     @GetMapping("/test/findAll")
